reorg theme selector
