+++ packages/ui/src/components/theme-selector-v2/index.ts
@@ -1,0 +1,2 @@
+export * from './theme-dialog'
+export * from './types'
