+++ packages/ui/src/components/theme-selector-v2/theme-dialog.tsx
@@ -3,53 +3,7 @@ import React, { FC, useEffect, useState } from 'react'
 import { Button, ButtonGroup, Dialog, Separator, Text } from '@/components'
 import { cn } from '@/utils/cn'
 
-export interface ThemeInterface {
-  mode: Mode
-  contrast: Contrast
-  colorAdjustment: ColorAdjustment
-  accentColor: AccentColor
-}
-
-export enum Mode {
-  Dark = 'Dark',
-  Light = 'Light'
-}
-
-export enum Contrast {
-  Default = 'Default',
-  HighContrast = 'High Contrast',
-  Dimmer = 'Dimmer'
-}
-
-export enum ColorAdjustment {
-  Default = 'Default',
-  Protanopia = 'Protanopia',
-  Deuteranopia = 'Deuteranopia',
-  Tritanopia = 'Tritanopia'
-}
-
-export enum AccentColor {
-  Yellow = 'yellow',
-  Red = 'red',
-  Pink = 'pink',
-  Purple = 'purple',
-  Blue = 'blue',
-  Cyan = 'cyan',
-  Orange = 'orange',
-  Brown = 'brown',
-  Green = 'green'
-}
-
-interface ThemeDialogProps {
-  defaultTheme?: ThemeInterface
-  theme?: ThemeInterface
-  open: boolean
-  onOpenChange: (open: boolean) => void
-  onChange: (language: ThemeInterface) => void
-  onSave: (language: ThemeInterface) => void
-  onCancel: () => void
-  children: React.ReactNode
-}
+import { AccentColor, ColorAdjustment, Contrast, Mode, ThemeDialogProps } from './types'
 
 const ThemeDialog: FC<ThemeDialogProps> = ({
   defaultTheme,
